pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel>=0.38"]
build-backend = "setuptools.build_meta"

[project]
name = "phonebook"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn>=0.29,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "jinja2>=3.1,<4.0",
    "httpx>=0.27,<1.0",
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.21,<1.0",
    "ruff>=0.4,<1.0",
]

[project.scripts]
phonebook = "phonebook.main:run"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
